pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = ""
authors = ["sheirand <eugene.osakovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
